+++ src/OTsk.java
@@ -1,8 +1,9 @@
+import java.math.BigInteger;
 
 public class OTsk {
-	public final int w1,w2,v1,v2;
+	public final BigInteger w1,w2,v1,v2;
 	
-	public OTsk(int w1, int w2, int v1, int v2){
+	public OTsk(BigInteger w1, BigInteger w2, BigInteger v1, BigInteger v2){
 		this.w1 = w1; 
 		this.w2 = w2; 
 		this.v1 = v1;
